OAK-10973 -- refactor guava collection and move test broken surrogate to CompressedDocumentPropertyStateTest
